FIX: TokenRequired navigate 훅 대신 컴포넌트로 대체
